+++ include/enjin2/scripting/bindings.hpp
@@ -431,7 +431,7 @@ private:
         SpriteSheet sheet;       ///< Sheet data (pointer to external pixel data)
         float       fps{8.0f};   ///< Playback rate in frames per second
         float       accumSec{0.0f}; ///< Accumulated seconds since last frame advance
-        uint8_t     frame{0};    ///< Current frame index
+        uint16_t    frame{0};    ///< Current frame index
         AnimMode    mode{AnimMode::Loop}; ///< Animation loop mode
         bool        forward{true};  ///< Ping-pong direction (true = forward)
         bool        done{false};    ///< Once mode: true when animation has completed
